20260915 변경 사항 1. 2026_GuestBook_1Team.cpp 파일 최종 수정본 업데이트 2. PenStyle.cpp select_color 변수 초기값 설정 3. draw.h draw.cpp 구조체 변수 자료형 변경 및 변경값 적용

+++ 2026_GuestBook_1Team/2026_GuestBook_1Team/Draw.h
@@ -52,7 +52,7 @@ struct DrawPointData
 {
 	POINT point; ///< 좌표를 저장하는 변수
 	ULONGLONG elapsed_time; ///< 시간을 저장하는 변수
-	int radious;
+	float radious;
 };
 /**
 * @brief 선의 데이터를 저장하기 위한 구조체
